requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
decorator==4.4.2
distlib==0.3.4
Django==3.2.12
djangorestframework==3.13.1
filelock==3.6.0
idna==3.3
imageio==2.16.1
imageio-ffmpeg==0.4.5
importlib-metadata==4.11.3
instaloader==4.8.4
moviepy==1.0.3
numpy==1.21.5
Pillow==9.0.1
platformdirs==2.5.1
proglog==0.1.9
pytz==2021.3
requests==2.27.0
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.2
tqdm==4.63.0
typing_extensions==4.1.1
urllib3==1.26.8
virtualenv==20.13.4
zipp==3.7.0